requirement.txt
altair @ file:///home/conda/feedstock_root/build_artifacts/altair-split_1734244716962/work
altgraph==0.17.5
annotated-types==0.7.0
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1759762205280/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///D:/bld/brotli-split_1761591996260/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1760287617600/work
certifi @ file:///C:/miniconda3/conda-bld/certifi_1763135349059/work/certifi
cffi @ file:///D:/bld/cffi_1761202893018/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click @ file:///D:/bld/click_1758270401364/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comtypes==1.4.13
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1735887193964/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1753444155689/work
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
googlemaps==4.10.0
grpcio==1.76.0
grpcio-status==1.71.2
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httplib2==0.31.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jsonschema_1755595646/work
jsonschema-specifications @ file:///tmp/tmptzik3dqa/src
MarkupSafe @ file:///D:/bld/markupsafe_1759055179018/work
narwhals @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_narwhals_1762277021/work
numpy==2.2.6
opencv-python==4.12.0.88
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///D:/bld/pandas_1759265558267/work
pefile==2023.2.7
pillow @ file:///D:/bld/bld/rattler-build_pillow_1758208682/work
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.11
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.12.4
pydantic_core==2.41.5
pydeck @ file:///home/conda/feedstock_root/build_artifacts/pydeck_1738346561094/work
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.10
pyparsing==3.2.5
pypiwin32==223
PySocks @ file:///D:/bld/pysocks_1733217287171/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.2.1
pyttsx3==2.99
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
pywin32==311
pywin32-ctypes==0.2.3
PyYAML @ file:///D:/bld/pyyaml_1758891577448/work
referencing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_referencing_1760379115/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
rpds-py @ file:///D:/bld/bld/rattler-build_rpds-py_1761178421/work
rsa==4.9.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1739781697784/work
SpeechRecognition==3.14.3
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1761776807357/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1743640740204/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tornado @ file:///D:/bld/tornado_1762506678249/work
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
uritemplate==4.2.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
watchdog @ file:///D:/bld/watchdog_1756135262508/work
win_inet_pton @ file:///D:/bld/win_inet_pton_1733130564612/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.25.0
